% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals}
\alias{animals}
\title{animals}
\format{
Un objeto de tipo data.frame.
}
\usage{
animals
}
\description{
Base de datos con cinco componentes de la leche materna de 25 especies de animales.
}
\keyword{datasets}
