% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apartamentos.R
\docType{data}
\name{apartamentos}
\alias{apartamentos}
\title{Datos del precio de 141 apartamento con su área}
\format{
Base de datos con 141 filas y 2 variables:
\describe{
\item{precio}{precio del apartamento en millones de pesos}
\item{area.cons}{área del apartamento}
}
}
\usage{
apartamentos
}
\description{
La base corresponde a datos sobre 141 apartamentos para fines academicos.
}
\keyword{datasets}
