% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arboles1.R
\docType{data}
\name{arboles1}
\alias{arboles1}
\title{Datos de 90 árboles}
\format{
Base de datos con 141 filas y 2 variables:
\describe{
\item{finca}{ubicación del árbol}
\item{mg}{genotipo del árbol}
\item{preso}{peso del árbol en ton}
\item{diametro}{diametro del árbol}
\item{altura}{altura del árbol}
}
}
\usage{
arboles1
}
\description{
La base corresponde a datos sobre 90 árboles
}
\keyword{datasets}
