% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditos.R
\docType{data}
\name{creditos}
\alias{creditos}
\title{Datos del precio de 780 clientes de un banco}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 780 rows and 5 columns.

base  con 780 filas y 5 variables:
\describe{
\item{default}{tiene el valor de 1 si el cliente no ha pagado y 0 si ha pagado su crédito}
\item{antiguedad}{antiguedad del cliente en el banco}
\item{edad}{edad del ciente}
\item{cuota}{cuota del crédito que debe pagar}
\item{ingresos}{ingresos del cliente}
}
}
\source{
base creada para fines académicos.
}
\usage{
creditos

creditos
}
\description{
La base corresponde a datos sobre 780 clientes de un banco con 5 variables.
}
\details{
La base contiene informacion sobre el comportamiento de 780 clientes de un banco y sus principales caracteristicas.
}
\examples{
data(creditos)
summary(creditos)

}
\keyword{datasets}
